+++ src/adapters/node.ts
@@ -4,7 +4,6 @@ import type {
   Server,
   ServerHandler,
   ServerOptions,
-  ServerRequest,
 } from "../types.ts";
 import NodeHttp from "node:http";
 import NodeHttps from "node:https";
@@ -25,7 +24,7 @@ export function toNodeHandler(fetchHandler: FetchHandler): NodeHttpHandler {
     nodeReq: NodeHttp.IncomingMessage,
     nodeRes: NodeHttp.ServerResponse,
   ) => {
-    const request = new NodeRequestProxy(nodeReq) as ServerRequest;
+    const request = new NodeRequestProxy(nodeReq);
     request.node = { req: nodeReq, res: nodeRes };
     const res = fetchHandler(request);
     return res instanceof Promise
@@ -58,7 +57,7 @@ class NodeServer implements Server {
       nodeReq: NodeHttp.IncomingMessage,
       nodeRes: NodeHttp.ServerResponse,
     ) => {
-      const request = new NodeRequestProxy(nodeReq) as ServerRequest;
+      const request = new NodeRequestProxy(nodeReq);
       request.node = { req: nodeReq, res: nodeRes };
       const res = fetchHandler(request);
       return res instanceof Promise